+++ package.json
@@ -10,11 +10,10 @@
   },
   "dependencies": {
     "@headlessui/react": "^2.2.4",
-    "@heroicons/react": "^2.2.0",
     "axios": "^1.10.0",
+    "next": "15.4.1",
     "react": "19.1.0",
-    "react-dom": "19.1.0",
-    "next": "15.4.1"
+    "react-dom": "19.1.0"
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3.3.1",
